Query context instead of platform to find out about double support in OpenCL Release command queue in OCLUtil instead of CContext.
There are two CContext instances in Source/core/rivertrail/jslib now.
We use the first instance to get the context extension,
and this instance is released soon.
Since the command queue is created only once at the very beginning of the process,
and it will be used by the second CContext instance later,
we cannot release the command queue in the destructor of the first CContext instance.