treat 'this' as unimplemented in type inference.
When we are in jslib, there's no way to know what 'this' refers to in the context of the caller. For example, let's say that 'arr.mapPar(function (x) {return this;})' is called in function 'Obj.f()'. In function 'mapPar', 'this' refers to 'arr', while in function 'f', 'this' refers to 'Obj'. Therefore, 'eval(this)' in jslib doesn't solve our problem.