Fix a bug in generating 'return' statements.
The original code generator will allocate memory in private space for arrays and inline objects returned by the called function, so the caller should view these return values as '__private'. Also, marking these return values as private are in accord with the fact that the memory for the returned array objects is allocated in a nested way, instead of the way for a flat global array. Thus, while generating the return statement in a called scope, the allocated memory for the return value should always be addressed with nested index operators, no matter whether the source is in private space or in global space.