Update runOCL.js.
The start address of the source and the result memory are passed to the kernel as arguments. If the kernel returns successfully, the result is converted to an Array object. The element of the result array could be either a nested array or an object. This commit also includes some code clean-up to remove the legacy ParallelArray-related literals.