WIP: Add support for global variable access.
In this patch, a new field 'extraBindings' is added to the type environment of a function. When a global is encountered for the first time in driving a function, there isn't a corresponding binding entry in the symbol table yet, so we evaluate the value of the global and add the inferred type to the 'extraBindings' field of the function's type environment. Thus, the 'extraBindings' field remembers all the global variables referenced in a function, and if a global is referenced again in this function, 'lookup' will find a matching entry in 'extraBindings'. It's really necessary to remember all the referenced globals in a function because we pass a global variable to the top-level kernel function or a called function as an argument. That is to say, besides the arguments / parameters appeared explicitly in the JavaScript code, some extra arguments / parameters will be added by the OpenCL code generator according to the callee's 'extraBindings' field. It should also be noted that if function 'f' calls function 'g', then all globals appeared in function 'g' should be merged by function 'f' so that the globals can be passed correctly to function 'g' from the caller of function 'f'.